Adds optional chaining. Moves query from device to project. Renders buttons only when query complete.
