+++ src/frontend/hooks/server/deviceInfo.ts
@@ -3,10 +3,8 @@ import {useMutation, useQuery, useQueryClient} from '@tanstack/react-query';
 import {deviceType, DeviceType} from 'expo-device';
 
 import {useApi} from '../../contexts/ApiContext';
-import {useActiveProject} from '../../contexts/ActiveProjectContext';
 
 export const DEVICE_INFO_KEY = 'deviceInfo';
-export const CREATED_BY_TO_DEVICE_ID_KEY = 'createdByToDeviceId';
 
 export const useDeviceInfo = () => {
   const mapeoApi = useApi();
@@ -33,17 +31,6 @@ export const useEditDeviceInfo = () => {
     },
     onSuccess: () => {
       queryClient.invalidateQueries({queryKey: [DEVICE_INFO_KEY]});
-    },
-  });
-};
-
-export const useCreatedByToDeviceId = (createdBy: string) => {
-  const project = useActiveProject();
-
-  return useQuery({
-    queryKey: [CREATED_BY_TO_DEVICE_ID_KEY, createdBy],
-    queryFn: async () => {
-      return await project.$createdByToDeviceId(createdBy);
     },
   });
 };
